chore: move all secp256k1 helpers to primitives-traits
